+++ CHANGELOG.rst
@@ -9,8 +9,9 @@ v2.2.0
 
 * Use the system's date format as a default.
 * Show "Today" or "Tomorrow" when due date is today/tomorrow respectively.
-* Add ``--start="[before|after] <dat>`` option for ``todo list``. It shows lists 
+* Add ``--start="[before|after] <date>`` option for ``todo list``. It shows todos
   created before/after given date.
+
 v2.1.0
 ------